+++ ron/bench/Main.hs
@@ -1,6 +1,10 @@
-{-# LANGUAGE NamedFieldPuns #-}
+{-# OPTIONS -Wno-orphans #-}
 
-import           Control.DeepSeq (force)
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE StandaloneDeriving #-}
+
+import           Control.DeepSeq (NFData, force)
 import           Control.Exception (evaluate)
 import           Control.Monad (void)
 import           Criterion (bench, nf)
@@ -8,8 +12,16 @@ import           Criterion.Main (defaultConfig, defaultMainWith)
 import           Criterion.Types (timeLimit)
 
 import           RON.Text (parseWireFrames, serializeWireFrames)
-import           RON.Types (Op (..), RawOp (..), WireChunk (Raw))
+import           RON.Types (Atom, Op (..), RawOp (..), UUID, WireChunk (Raw),
+                            WireReducedChunk)
 import qualified RON.UUID as UUID
+
+deriving instance NFData Atom
+deriving instance NFData Op
+deriving instance NFData RawOp
+deriving instance NFData UUID
+deriving instance NFData WireChunk
+deriving instance NFData WireReducedChunk
 
 main :: IO ()
 main = do